Add registration for 2021 Cloud Engineering Summit (#562)
* First cut of sessionize integration

* Add working registration flow

* Some copy updates and formatting fixes

* CES registration updates

* Address some feedback

* Remove uneeded pages and update ctas

* Remove unused image

* Address feedback

+++ pulumi-hugo/themes/default/components/src/components/datetime/datetime.tsx
@@ -17,7 +17,7 @@ export class Datetime {
         const date = new Date(this.date);
 
         const options: Intl.DateTimeFormatOptions = {
-            timeZoneName: "short",
+            timeZoneName: "long",
             weekday: "short",
             year: "numeric",
             month: "long",
